docker/requirements.txt
gunicorn==20.0.4
asgiref==3.4.1
backcall==0.2.0
dataclasses
decorator==5.0.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
whitenoise
bcrypt==3.2.2
# drf_yasg
# requests
# bson
# pymongo
# pyjwt
# django-rest-swagger
# django_mongodb_engine
Django==3.0.5
django_mongodb_engine==0.6.0
django_rest_swagger==2.2.0
djangorestframework==3.13.1
djongo==1.3.3
drf_yasg==1.20.0
firebase_admin==5.2.0
pymongo==3.12.0
# python_bcrypt==0.3.2
pyjwt